requirements.txt
alembic==1.5.8
autopep8==1.5.6
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.5
click==7.1.2
fastapi==0.63.0
greenlet==1.0.0
h11==0.12.0
httpcore==0.12.3
httpx==0.17.1
idna==3.1
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
passlib==1.7.4
psycopg2-binary==2.8.6
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.8.1
PyJWT==2.0.1
python-dateutil==2.8.1
python-editor==1.0.4
rfc3986==1.4.0
six==1.15.0
sniffio==1.2.0
SQLAlchemy==1.4.11
starlette==0.13.6
toml==0.10.2
typing-extensions==3.7.4.3
uvicorn==0.13.4
